requirements.txt
pyserial
paho-mqtt
pandas
